fix plugin test so that the test plugin isn't packaged w/ types
