+++ types/video.js/tsconfig.json
@@ -19,7 +19,7 @@
     },
     "files": [
         "index.d.ts",
-        "video.js-some-plugin.d.ts",
+        "test/plugin-tests.ts",
         "video.js-tests.ts"
     ]
 }
