+++ types/video.js/video.js-tests.ts
@@ -69,8 +69,6 @@ videojs("example_video_1").ready(function() {
 
 	testPlugin(this, {});
 
-	testAugmentation(this);
-
 	testLogger();
 });
 
@@ -118,21 +116,6 @@ function testPlugin(player: videojs.Player, options: {}) {
 		});
 	});
 	(player as any).uloztoExample(options);
-}
-
-function testAugmentation(player: videojs.Player) {
-	player.somePluginDefinedInAugmentation();
-	videojs('example_video_2', {
-        plugins: {
-            somePluginDefinedInAugmentation: {
-                someRequiredProperty: true
-            },
-            someOtherPluginNotTyped: {},
-        },
-    });
-	videojs('example_video_3', {
-        plugins: {}
-    });
 }
 
 function testLogger() {
